prom-mcp-server/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
uvicorn[standard]>=0.30.0
starlette>=0.40.0
python-dotenv>=1.0.0
